Reduced the size of num_moves in the move_stack_t.
git-svn-id: file:///home/shlomif/Backup/svn-dumps/google-code/svnsync-repos/fc-solve/trunk@3258 e7e8a897-7ba4-4ee7-b36f-f4c66519b19a

+++ fc-solve/source/internal_move_struct.h
@@ -96,7 +96,7 @@ typedef fcs_move_t fcs_internal_move_t;
 typedef struct
 {
     fcs_internal_move_t * moves;
-    int num_moves;
+    unsigned short num_moves;
 } fcs_move_stack_t;
 
 #ifdef __cplusplus
